pyproject.toml
[project]
name = "recommender"
version = "0.1.0"
dependencies = [
    "torch",
    "torch-geometric",
    "pandas",
    "numpy",
    "requests",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
]

[tool.setuptools.packages.find]
where = ["src"]